(isla_machinery/nightly_ifetch) add a fixed pc limit for ifetch tests
Discard traces with repeated PCs, allowing the test to succeed
(with allow/forbid) on the traces that actually terminated.

There is no 'safe' choice for a pc limit, and eventually we want
isla to report tests with no satisfiable executions but with discarded
traces as "indeterminate" rather than just "forbidden" in future.